\name{IlluminaHumanMethylationEPICanno.ilm10b4.hg19}
\alias{IlluminaHumanMethylationEPICanno.ilm10b4.hg19}
\alias{Locations}
\alias{Other}
\alias{SNPs.Illumina}
\alias{Manifest}
\alias{Islands.UCSC}
\docType{data}
\title{
  Annotation data for the \sQuote{IlluminaHumanMethylationEPIC} micro array.
}
\description{
 This package is based on the file
  \sQuote{MethylationEPIC_v-1-0_B4.csv} from \url{www.illumina.com}.

  Additional SNP annotation is generated by the authors (described below).

  A script for creating the data object in this package is contained in
  \code{scripts/manifest.R}.
}
\format{
  An object of class \code{IlluminaMethylationAnnotation} for
  \code{IlluminaHumanMethylationEPICanno.ilm10b4.hg19}. All others are of
  class \code{DataFrame}
}
\details{
  The following changes/ modifications / addition has been made to the
  source material.

  For the annotation related to UCSC islands, a value of \code{""} has
  been changed to \code{"OpenSea"}.

  The creation of this object based on the Illumina annotation and
  additional SNP information (see below) is contained in the
  \code{createAnno.R} script in the \code{scripts} directory.

  Contents of the columns are the responsibility of Illumina; see their
  documentation for details.
}

\source{
  See description.
}

\examples{
data(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
data(Locations)
data(Other)
data(Manifest)
data(SNPs.Illumina)
data(Islands.UCSC)
}
\keyword{datasets}
